% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_census.R
\name{c11_ethdetail_ew}
\alias{c11_ethdetail_ew}
\title{Get 2011 Census data from Nomis: detailed ethnic groups, England & Wales}
\usage{
c11_ethdetail_ew(geog, output = "n", summarise = FALSE)
}
\arguments{
\item{geog}{A string or vector of strings specifying geographic areas using ONS nine-digit codes, (e.g "E07000178")}

\item{output}{Specify "n" for counts or "p" for percentages. Defaults to "n".}

\item{summarise}{If TRUE, creates a single row for each instance of a detailed
ethnic group, even if that detailed group is used across multiple broad groups
(e.g 'White: Afghan' and 'Asian: Afghan' totals are added together into a
single 'Afghan' row). If FALSE, keeps instances of single ethnic groups separate
if they are reported in separate broad groups (e.g. 'White: Afghan' and 'Asian:
Afgan' are reported separately). Defaults to FALSE.}
}
\value{
A tibble with a row for each detailed ethnic group for each geographic
area specified, and a column showing the number or percentage as specified.
If summarise is FALSE, a further column shows the broad ethnic group.
}
\description{
Nomis data retrieval function. Returns data from the 2011 Census on
population by detailed ethnic group. Note that outputs with 'and' in the
name refer to people who identify with Mixed or Multiple ethnic groups.
}
\examples{

eth_pop_full <- c11_ethgrps_ew(geog = c("E06000014", "E07000121"))

eth_percent_summary <- c11_ethgrps_ew(geog = c("E06000014", "E07000121"), output = "p", summarise = TRUE)

}
