% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_census.R
\name{c11_ethgrps_uk}
\alias{c11_ethgrps_uk}
\title{Get 2011 Census data from Nomis: broad ethnic groups, UK}
\usage{
c11_ethgrps_uk(geog, output = "n")
}
\arguments{
\item{geog}{A string or vector of strings specifying geographic areas using ONS nine-digit codes, (e.g "E07000178")}

\item{output}{Specify "n" for counts or "p" for percentages. Defaults to "n".}
}
\value{
A tibble with a row for each geographic area specified, and a column
for each broad ethnic group.
}
\description{
Nomis data retrieval function. Returns data from the 2011 Census on
population by ethnic group, using the broad ethnic group classifications
used for geographic areas across the UK.
}
\examples{

Does this show up as commentary text in the example?

eth_pop <- c11_ethgrps_uk(geog = c("S12000036", "S12000046"))

eth_percent <- c11_ethgrps_uk(geog = c("S12000036", "S12000046"), output = "p")

}
