% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_census.R
\name{c11_tenure_uk}
\alias{c11_tenure_uk}
\title{Get 2011 Census data from Nomis: housing tenure, UK}
\usage{
c11_tenure_uk(geog, output = "n")
}
\arguments{
\item{geog}{A string or vector of strings specifying geographic areas using ONS nine-digit codes, (e.g "E07000178")}

\item{output}{Specify "n" for counts or "p" for percentages. Defaults to "n".}
}
\value{
A tibble with a row for each geographic area specified, and columns
    for each tenure group.
}
\description{
Nomis data retrieval function. Returns data from the 2011 Census on
housing tenure (i.e home ownership and social/private renting). Works for
all available Census geographies in the UK.
}
\examples{

tenure_pop <- c11_tenure_uk(geog = c("E06000014", "E07000121"))

cob_percent <- c11_tenure_uk(geog = c("E06000014", "E07000121"), output = "p")

}
