% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\docType{data}
\name{census_menu}
\alias{census_menu}
\title{Key: list of 2011 Census tables that the "get_c11..." functions in this package can work with.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 96 rows and 3 columns.
}
\source{
ONS
}
\usage{
census_menu
}
\description{
This list shows all UK-based tables available from Nomis, as well as
most of the Key Statistics series for England and Wales.
The get_c11_uk and get_c11_ew functions should work for most of these
tables - no guarantees though.
}
\keyword{datasets}
