% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data.R
\docType{data}
\name{con2rn}
\alias{con2rn}
\title{A dataframe showing a constituency to region/nation lookup}
\format{
A data frame with 650 observations and 4 variables.
\describe{
\item{con_code}{9-digit ONS code of the constituency}
\item{con_name}{the name of the constituency}
\item{rn_code}{the 9-digit ONS code of the region/nation the constituency is in}
\item{rn_name}{the name of the region/nation the constituency is in}
}
}
\source{
ONS
}
\usage{
con2rn
}
\description{
A dataframe showing the name and ONS 9-digit code of each constituency in
the UK, along with the name and 9-digit code of the English region the
constituency is in (if in England) or the nation (if it's in a devolved
nation).
}
\keyword{datasets}
