% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_census.R
\name{find_censustable}
\alias{find_censustable}
\title{Find the title of a 2011 Census table}
\usage{
find_censustable(title)
}
\arguments{
\item{title}{A string containing the ONS title of a 2011 Census table.}
}
\value{
A string containing the ID used to identify that table in Nomis.
}
\description{
Returns the Nomis ID of a 2011 Census table based on its ONS title, e.g. 'KS201EW'.
This makes it easier to use nomisr's \code{nomis_get_data} function for census tables.
}
\examples{
tenure_id <- find_censustable("KS402UK")

}
