% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_age.R
\name{get_age}
\alias{get_age}
\title{Retrieve population by age data from Nomis}
\usage{
get_age(geog, sex = "t", year = "latest", output = "n")
}
\arguments{
\item{geog}{A string or vector of strings specifying geographic areas using ONS nine-digit codes, (e.g "E07000178")}

\item{sex}{Specify "m" for male, "f" for female, or "t" for total. Defaults to "t".}

\item{year}{An integer specifying the year for which data is needed, or use the string "latest" for the most recent year available. Defaults to "latest".}

\item{output}{Specify "n" for counts or "p" for percentages. Defaults to "n".}
}
\value{
A tibble with all available age data from Nomis.
}
\description{
Returns raw population by age data from Nomis for specified sex(es),
geography(ies) and year. Determines automatically whether data should
be retrieved from the small-area-based estimates or the local-authority-
based estimates.
}
\details{
Local-authority based estimates (for LAs and any larger geographies) are
available from 1991 across the UK. Small-area based estimates (for any
geography smaller than an LA) are available from 2011 for England and Wales.
}
\examples{
oxford_raw <- get_age(geog = "E07000178")

}
