% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_census.R
\name{get_c11_uk}
\alias{get_c11_uk}
\title{Get 2011 Census data from Nomis: UK-wide tables}
\usage{
get_c11_uk(title, geog, output = "n")
}
\arguments{
\item{title}{A string containing the ONS title of a 2011 Census table.}

\item{geog}{A string or vector of strings specifying geographic areas using ONS nine-digit codes, (e.g "E07000178")}

\item{output}{Specify "n" for counts or "p" for percentages. Defaults to "n".}
}
\value{
A tibble with a row for each geographic area specified, and columns
    for each of the values within the demographic variable covered by the
    requested table.
}
\description{
Nomis data retreival function. Returns data for some 2011 Census tables if they
are designated as UK tables, rather than England & Wales tables. It only works
for single-variable tables, and isn't guaranteed to work for all of them.
}
\details{
See the census_menu dataset in this package for a list of census tables to try.
UK tables are those with a '...UK' suffix.
}
\examples{
oxford_cars <- get_c11_uk(title = "KS404UK", geog = "E07000178")

hours_percents <- get_c11_uk(title = "KS604UK", geog = c("S12000036", "S12000046"), output = "p")

}
