% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookups.R
\name{get_msoanames}
\alias{get_msoanames}
\title{Get a dataframe with the latest Commons Library MSOA names}
\usage{
get_msoanames()
}
\value{
A dataframe with ONS 9-digit code, ONS name in English and Welsh, Commons Library name in English and Welsh (Welsh MSOAs only), and LA name.
}
\description{
Returns a dataframe that shows, for each MSOA in England and Wales, the ONS
name and code for the MSOA, the Commons Library name, and the local authority
that the MSOA name is in. Data is retrieved from visual.parliament.uk.
}
\examples{
get_msoanames()

}
