% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookups.R
\name{get_oa2bua}
\alias{get_oa2bua}
\title{Get an output area to Built Up Area lookup (search by multiple Built Up Areas)}
\usage{
get_oa2bua(buas)
}
\arguments{
\item{buas}{A string, or a vector of strings, specifying the 9-digit ONS code for your chosen Built Up Area(s)}
}
\value{
A dataframe with names/codes for output areas, and the names/codes of associated Built Up Areas for each one.
}
\description{
Retrieves a dataframe from the ONS Geoportal API showing an output area to
Built Up Area lookup for a specified set of Built Up Areas (2011). The
lookup for all Output Areas is too large to return in one go, which is why
this function asks you to search by Built Up Area.
}
\details{
England and Wales only.
}
\examples{
get_oa2bua(buas = c("E34001405", "E34002311", "E34000732"))

}
