% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookups.R
\name{get_oa2bua_single}
\alias{get_oa2bua_single}
\title{Get an output area to Built Up Area lookup (search by single Built Up Area)}
\usage{
get_oa2bua_single(bua)
}
\arguments{
\item{bua}{A string specifying the 9-digit ONS code for a single Built Up Area}
}
\value{
A dataframe with names/codes for output areas, and the names/codes of the associated Built Up Area
}
\description{
Retrieves a dataframe from the ONS Geoportal API showing an Output Area to
Built Up Area lookup for a specified Built Up Area (2011) and year. This function
feeds into the more useful get_bua2ward function, which lets you retrieve
data for multiple Built Up Areas at once.
}
\details{
England and Wales only.
}
\examples{
get_oa2bua_single(bua = "E34004922")

}
