% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookups.R
\name{get_oa2buasd}
\alias{get_oa2buasd}
\title{Get an output area to Built Up Area Subdivision lookup (search by multiple Built Up Area Subdivisions)}
\usage{
get_oa2buasd(buasds)
}
\arguments{
\item{buasds}{A string, or a vector of strings, specifying the 9-digit ONS code for your chosen Built Up Area Subdivision(s)}
}
\value{
A dataframe with names/codes for output areas, and the names/codes of associated Built Up Area Subdivisions and Built Up Areas.
}
\description{
Retrieves a dataframe from the ONS Geoportal API showing an output area to
Built Up Area Subdivision (BUASD) to Built Up Area lookup for a specified set
of BUASDs (2011). Thelookup for all Output Areas is too large to return in one
go, which is why this function asks you to search by BUASD.
}
\details{
England and Wales only.
}
\examples{
get_oa2buasd(buasds = c("E35000528", "E35000500", "E35001362", "E35001381"))

}
