% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookups.R
\name{get_oa2buasd_single}
\alias{get_oa2buasd_single}
\title{Get an output area to Built Up Area Subdivision lookup (search by single Built Up Area Subdivision)}
\usage{
get_oa2buasd_single(buasd)
}
\arguments{
\item{buasd}{A string specifying the 9-digit ONS code for a single Built Up Area Subdivision}
}
\value{
A dataframe with names/codes for output areas, and the names/codes of the associated Built Up Area Subdivision and Built Up Area
}
\description{
Retrieves a dataframe from the ONS Geoportal API showing an Output Area to
Built Up Area Subdivision (BUASD) lookup for a specified BUASD (2011) and year. This function
feeds into the more useful get_buasd2ward function, which lets you retrieve
data for multiple BUASDs at once.
}
\details{
England and Wales only.
}
\examples{
get_oa2buasd_single(buasd = "E35001205")

}
