% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookups.R
\name{get_oa2ward}
\alias{get_oa2ward}
\title{Get an output area to ward and local authority lookup (search by multiple wards)}
\usage{
get_oa2ward(wards, year)
}
\arguments{
\item{wards}{A string, or a vector of strings, specifying the 9-digit ONS code for your chosen ward(s)}

\item{year}{A number for the desired year, in format YYYY}
}
\value{
A dataframe with names/codes for output areas, and the names/codes of associated wards and local authorities for each one.
}
\description{
Retrieves a dataframe from the ONS Geoportal API showing an output area to
ward to local authority lookup for a specified set of wards and a year. The
lookup for all Output Areas is too large to return in one go, which is why
this function asks you to search by ward.
}
\details{
Output areas are matched to wards on a best-fit basis. England and Wales only.
}
\examples{
get_oa2ward(ward = c("E05010317", "E05010319", "E05010326"), year = 2019)

}
