% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookups.R
\name{get_oa2ward_single}
\alias{get_oa2ward_single}
\title{Get an output area to ward and local authority lookup (search by single ward)}
\usage{
get_oa2ward_single(ward, year)
}
\arguments{
\item{ward}{A string specifying the 9-digit ONS code for a single ward}

\item{year}{A number for the desired year, in format YYYY}
}
\value{
A dataframe with names/codes for output areas, and the names/codes of the associated ward and local authority.
}
\description{
Retrieves a dataframe from the ONS Geoportal API showing an Output Area to
ward to local authority lookup for a specified ward and year. This function
feeds into the more useful get_oa2ward function, which lets you retrieve
data for multiple wards at once.
}
\details{
Output areas are matched to wards on a best-fit basis. England and Wales only.
}
\examples{
get_oa2ward_single(ward = "E05010317", year = 2019)

}
