% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lookups.R
\name{get_ward2con}
\alias{get_ward2con}
\title{Get a ward to parliamentary constituency lookup (search by single constituency)}
\usage{
get_ward2con(con, year)
}
\arguments{
\item{con}{A string specifying the 9-digit ONS code for your chosen constituency}

\item{year}{A number for the desired year, in format YYYY}
}
\value{
A dataframe with names/codes for wards, and the names/codes for associated local authorities for each one.
}
\description{
Retrieves a dataframe from the ONS Geoportal API showing a ward to constituency
lookup for a specified UK parliamentary constituency and year.
}
\details{
At time of writing, data is available through the API for 2016 through to 2020.
2021 data is published on the ONS Geoportal website, but not yet available
via the API.
}
\examples{
get_ward2con(la = "E14000639", year = 2019)

}
