% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data.R
\docType{data}
\name{la_changes}
\alias{la_changes}
\title{A dataframe showing district-level local authority changes in England (2019-2021)}
\format{
A data frame with 25 observations and 7 variables.
\describe{
\item{oa_code}{9-digit ONS code of a 2011 Census Output Area}
\item{con_code}{9-digit ONS code of the UK parliamentary constituency the Output Area best-fits to}
\item{con_name}{name of the UK parliamentary constituency the Output Area best-fits to}
}
}
\source{
Commons Library analysis of ONS Open Geography datasets
}
\usage{
la_changes
}
\description{
A dataframe showing which former district local authorities were combined to
create new district local authorities in 2019, 2020, and 2021. Also shows
which region each of these local authorities are in.
}
\keyword{datasets}
