% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_age.R
\name{pop_deciles}
\alias{pop_deciles}
\title{Retrieve population deciles for a given area from Nomis}
\usage{
pop_deciles(geog, sex = "t", year = "latest", output = "n")
}
\arguments{
\item{geog}{A string or vector of strings specifying geographic areas using ONS nine-digit codes, (e.g "E07000178")}

\item{sex}{Specify "m" for male, "f" for female, or "t" for total. Defaults to "t".}

\item{year}{An integer specifying the year for which data is needed, or use the string "latest" for the most recent year available. Defaults to "latest".}

\item{output}{Specify "n" for counts or "p" for percentages. Defaults to "n".}
}
\value{
A tibble with a row for each geography specified and columns for the year and each age band.
}
\description{
Returns the population in ten-year age bands for a given set of geographies
with sex and year speified. Automatically retrieves data from
small-area-based and/or local-authority-based estimates as appropriate.
}
\examples{

oxford_deciles <- pop_deciles(geog = "E07000178")

oxford_men_percent <- pop_deciles(geog = "E07000178", sex = "m", output = "p")

}
