% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_age.R
\name{pop_range}
\alias{pop_range}
\title{Retrieve population within a specified age range for a given area from Nomis}
\usage{
pop_range(
  geog,
  sex = "t",
  year = "latest",
  lower = 0,
  upper = 90,
  output = "n"
)
}
\arguments{
\item{geog}{A string or vector of strings specifying geographic areas using ONS nine-digit codes, (e.g "E07000178")}

\item{sex}{Specify "m" for male, "f" for female, or "t" for total. Defaults to "t".}

\item{year}{An integer specifying the year for which data is needed, or use the string "latest" for the most recent year available. Defaults to "latest".}

\item{lower}{An integer specifying the lower bound of the age range. Defaults to 0.}

\item{upper}{An integer specifying the upper bound of the age range. Defaults to 90 (the integer for all ages 90 and above).}

\item{output}{Specify "n" for counts or "p" for percentages. Defaults to "n".}
}
\value{
A tibble with a row for each geography specified.
}
\description{
Returns the population for a given age range and a given set of geographies
with sex and year speified. Automatically retrieves data from
small-area-based and/or local-authority-based estimates as appropriate.
Note that in this dataset, the integer '90' stands in for all ages from 90
upwards.
}
\examples{

oxford_children_f <- pop_range(geog = "E07000178", sex = "f", lower = 0, upper = 17)

oxford_60_plus_pc <- pop_range(geog = "E07000178", lower = 60)

}
