% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomis_age.R
\name{pop_total}
\alias{pop_total}
\title{Retrieve total population for a given area from Nomis}
\usage{
pop_total(geog, sex = "t", year = "latest")
}
\arguments{
\item{geog}{A string or vector of strings specifying geographic areas using ONS nine-digit codes, (e.g "E07000178")}

\item{sex}{Specify "m" for male, "f" for female, or "t" for total. Defaults to "t".}

\item{year}{An integer specifying the year for which data is needed, or use the string "latest" for the most recent year available. Defaults to "latest".}
}
\value{
A tibble with a row for each geography selected and columns showing the year and total population.
}
\description{
Returns the total population for a given set of geographies with sex
and year speified. Automatically retrieves data from small-area-based
and/or local-authority-based estimates as appropriate.
}
\examples{
oxford_total <- pop_total(geog = "E07000178")

oxford_women <- pop_total(geog = "E07000178", sex = "f", year = 2019)

}
